Whl sampling rate configurable in PlaceFieldProcessor, default 512 Float whl sampling rate in WhlFileReader

+++ sdl_example/WhlFileReaderProcessor.cpp
@@ -10,7 +10,7 @@
 WhlFileReaderProcessor::WhlFileReaderProcessor(LFPBuffer* buffer)
 : WhlFileReaderProcessor(buffer,
 		buffer->config_->getString("dat.path.base") + buffer->config_->getString("whl.extension", "whl"),
-		buffer->config_->getInt("whl.sampling.factor"),
+		buffer->config_->getFloat("whl.sampling.factor"),
 		buffer->config_->getFloat("whl.sub.x"),
 		buffer->config_->getFloat("whl.sub.y")
 ) {
